wxGUI: icon size is float, not int (#1035)
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@52915 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ gui/wxpython/mapdisp/frame.py
@@ -882,7 +882,7 @@ class MapFrame(SingleMapFrame):
             if '=' in p:
                 parg,pval = p.split('=')[0:2]
                 if parg == 'icon': icon = pval
-                elif parg == 'size': size = int(pval)
+                elif parg == 'size': size = float(pval)
 
         pattern = ["d.vect",
                    "map=%s" % name,
